:recycle: api: Move delete-unlinked-certification-candidate injection to session
